fixed order module
fixed order module
cleaned up index.js => store
