+++ server/app.js
@@ -12,11 +12,12 @@ const userRouter = require("./routes/users");
 
 const PORT = 5000;
 
-app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(cors());
 app.use(express.json());
+
 app.use("/user", userRouter);
 app.use("/api/beans", beansRouter);
 app.use("/shoppingcart", cartRouter);
@@ -32,10 +33,9 @@ mongoose
     }
   )
   .then(() => {
-    console.log("connected to databaase");
+    console.log(`Connected to MongoDB Airbean`);
+    app.listen(5000, () => console.log(`Server started on port ${PORT}`));
   })
-  .catch(error => {
-    console.log(`ERROR: ${error}`);
+  .catch(() => {
+    console.log(`Error connecting to the database`);
   });
-
-app.listen(5000, () => console.log(`Server started on port ${PORT}`));
