--- server/data/menu copy.json
@@ -1,46 +1,0 @@
-{
-  "menu": [
-    {
-      "product_id": 1,
-      "title": "Bryggkaffe",
-      "desc": "Bryggd på månadens bönor.",
-      "price": 39,
-      "quantity": 0
-    },
-    {
-      "product_id": 2,
-      "title": "Caffè Doppio",
-      "desc": "Bryggd på månadens bönor.",
-      "price": 49,
-      "quantity": 0
-    },
-    {
-      "product_id": 3,
-      "title": "Cappuccino",
-      "desc": "Bryggd på månadens bönor.",
-      "price": 49,
-      "quantity": 0
-    },
-    {
-      "product_id": 4,
-      "title": "Latte Macchiato",
-      "desc": "Bryggd på månadens bönor.",
-      "price": 49,
-      "quantity": 0
-    },
-    {
-      "product_id": 5,
-      "title": "Kaffe Latte",
-      "desc": "Bryggd på månadens bönor.",
-      "price": 54,
-      "quantity": 0
-    },
-    {
-      "product_id": 6,
-      "title": "Cortado",
-      "desc": "Bryggd på månadens bönor.",
-      "price": 39,
-      "quantity": 0
-    }
-  ]
-}
