+++ src/styles/global.scss
@@ -1,2 +1,3 @@
 @import "variables";
 @import "buttons";
+@import "transitions";
